+++ src/default-xrp-client.ts
@@ -3,8 +3,8 @@ import bigInt, { BigInteger } from 'big-integer'
 import { XRPClientDecorator } from './xrp-client-decorator'
 import TransactionStatus from './transaction-status'
 import RawTransactionStatus from './raw-transaction-status'
-import GRPCNetworkClient from './grpc-network-client'
-import GRPCNetworkClientWeb from './grpc-network-client.web'
+import GRPCNetworkClient from './grpc-xrp-network-client'
+import GRPCNetworkClientWeb from './grpc-xrp-network-client.web'
 import { NetworkClient } from './network-client'
 import { GetFeeResponse } from './generated/web/org/xrpl/rpc/v1/get_fee_pb'
 import { AccountAddress } from './generated/web/org/xrpl/rpc/v1/account_pb'
